Fix the gpu rendering problem

+++ RayTracing/src/Geometryrepository.cu
@@ -18,7 +18,7 @@ namespace Geometryrepository {
 	void Sendtogpu() {
 	}
 
-	__host__ __device__ Geometry* GetGeo(uint32_t geoid) {
+	__host__ __device__ Geometry* GetGeo(size_t geoid) {
 	#ifdef __CUDA_ARCH__
 		// Device
 		return g_Geos[geoid];
