+++ RayTracing/src/Geometryrepository.h
@@ -8,5 +8,5 @@ namespace Geometryrepository {
 	void Initiate(World const& world);
 	void Clearmemory();
 	void Sendtogpu();
-	__host__ __device__ Geometry* GetGeo(uint32_t geoid);
+	__host__ __device__ Geometry* GetGeo(size_t geoid);
 }
